requirements.txt
asgiref==3.2.10
Django==3.0.5
django-cors-headers==3.4.0
djangorestframework==3.11.1
djongo==1.3.3
joblib==0.16.0
mongoengine==0.20.0
numpy==1.19.1
pandas==1.1.0
patsy==0.5.1
pymongo==3.11.0
python-dateutil==2.8.1
pytz==2020.1
scikit-learn==0.23.2
scipy==1.5.2
six==1.15.0
sklearn==0.0
sqlparse==0.2.4
statsmodels==0.11.1
threadpoolctl==2.1.0
xgboost==1.1.1
